+++ components/market/myLink/announcement.tsx
@@ -112,8 +112,8 @@ const Announcement = memo(function Announcement({ data }: AnnouncementProps) {
         <button
           type="button"
           className={styles.Announcementsheader}
-          onClick={handleToggle}
-          onKeyDown={handleToggleKeys}
+          // onClick={handleToggle}
+          // onKeyDown={handleToggleKeys}
           aria-controls={contentId}
           aria-expanded={isExpanded}
           aria-label={announcementTitle}>