+++ i18n/en.ts
@@ -2476,7 +2476,6 @@ export default {
     SettingGeneral_Description: "Description",
     SettingGeneral_Send: "Send",
     // 6 notifications
-    SharedRemainingTimeBefore: "You should wait 23:54:34 to reshare again.",
     // #endregion Setting Section
     // // #region userpanel Section
     userpanel_market: "Market",
@@ -3445,5 +3444,10 @@ export default {
     Yes: "Yes",
     No: "No",
     // #endregion Notification Section
+
+    how: "How",
+    SharedRemainingTimeBefore: "You should wait 24 hours before resharing again.",
+    economy: "Economy",
+    Retry: "Retry",
   },
 };