+++ i18n/tr.ts
@@ -3438,5 +3438,14 @@ export default {
     Yes: "Evet",
     No: "Hayır",
     // #endregion Notification Section
+    how: "Nasıl",
+    SharedRemainingTimeBefore: "Tekrar paylaşım yapmadan önce 24 saat beklemelisiniz.",
+    messagesetting_ConnectProductResponseExplain: "Ürününüzü Story veya Live'a bağlayın.",
+    marketProperties_Connect: "Bağla",
+    economy: "Ekonomik",
+    Retry: "Tekrar Dene",
+    messagesetting_SpecifyProductResponseExplain:
+      "Ürün yanıtınızı belirleyin; böylece kullanıcılara otomatik olarak gönderilir.",
+    Storeproduct_NationalCardexplain: "Bu kart, cep telefonu numarasını Brancy'ye kaydeden kişiye ait olmalıdır.",
   },
 };